requirements.txt
pandas==2.2.1
requests==2.31.0
great_expectations==0.18.12
schedule==1.2.1
PyYAML==6.0.1
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.9
